refactor: move CLI launch behavior into runtime
